requirements.txt
absl-py==2.2.2
antlr4-python3-runtime==4.9.3
attrs==25.3.0
cffi==1.17.1
charset-normalizer==3.4.1
contourpy==1.3.0
cycler==0.12.1
dlib==19.24.8
face-alignment==1.4.1
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.57.0
fsspec==2025.3.2
huggingface-hub==0.30.2
idna==3.10
imageio==2.37.0
importlib-metadata==8.6.1
importlib-resources==6.5.2
jax==0.4.30
jaxlib==0.4.30
jinja2==3.1.6
kiwisolver==1.4.7
lazy-loader==0.4
llvmlite==0.43.0
markupsafe==3.0.2
matplotlib==3.9.4
mediapipe==0.10.21
ml-dtypes==0.5.1
mpmath==1.3.0
networkx==3.2.1
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
omegaconf==2.3.0
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opt-einsum==3.4.0
packaging==25.0
pillow==11.2.1
protobuf==4.25.6
pycparser==2.22
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pyyaml==6.0.2
requests==2.32.3
safetensors==0.5.3
scikit-image==0.24.0
scipy==1.13.1
sentencepiece==0.2.0
six==1.17.0
sounddevice==0.5.1
sympy==1.13.1
tifffile==2024.8.30
timm==1.0.15
torch==2.6.0
torch-summary==1.4.5
torchinfo==1.8.0
torchvision==0.21.0
tqdm==4.67.1
triton==3.2.0
typing-extensions==4.13.2
urllib3==2.4.0
zipp==3.21.0